+++ docs/src/releases.md
@@ -2,6 +2,7 @@
 ### 0.2.x
 - New Examples:
     - UserLib: how to build and call a user custom library providing additional Geant4 functionally that is not provided by the set of wrapped classes
+    - JuliaAction: emmbeding Julia in a C++ application. In this example we call user actions implemented in Julia
 
 ### 0.2.0
 - New Features